requirements.txt
Django==2.2.4
djangorestframework==3.10.3
pysmb==1.1.27
requests==2.22.0
django-pivot==1.8.1
